requirements.txt
Django~=1.11.4
psycopg2~=2.7.3
whitenoise~=3.3.0
django-crispy-forms~=1.6.1
django_guardian~=1.4.9
django-reversion~=2.0.10

Sickle~=0.6.2
surt~=0.3.0

uWSGI~=2.0.15
